==> code/01-preprocessing.ipynb <==
import os
import pandas as pd
import numpy as np
from pathlib import Path
from tqdm.notebook import tqdm # https://github.com/tqdm/tqdm
tqdm.pandas()

## 01 - Preprocessing

In this notebook process the raw files, which can be found in the `hhblock_dataset` folder. Each row in a block has a date, and along the columns, half-hourly blocks (compact form). We reshape this into long form so that each row has a date and a single half-hourly measurement (expanded form). We then combine the energy data with weather and bank holiday data to aid downstream analysis.

First, we read read and combine the rows from all blocks into a single dataframe:

source_data = Path('../data/')
block_data_path = source_data/'hhblock_dataset'

Some LCLid columns in the dataset end much earlier than the rest. This could be because particular households dropped out of the program early, moved house (leaving the house unoccupied) and so on. Because of this, We need to find the global latest date across all data blocks:

max_date = None
for f in tqdm(block_data_path.glob('*.csv')):
    df = pd.read_csv(f, parse_dates=False)
    df['day'] = pd.to_datetime(df['day'], yearfirst=True)
    if max_date is None:
        max_date = df['day'].max()
    else:
        if df['day'].max()>max_date:
            max_date = df['day'].max()
print(f'Max Date across all blocks: {max_date}')
del df

We define a function for converting the time series data into compact form:

def preprocess_compact(x, max_date):
    """Takes a DataFrame with time series data in compact form and a maximum date. 
    Returns the time series array, along with the time series identifier, start date,
    and the length of the series."""

    start_date = x['day'].min()
    name = x['LCLid'].unique()[0]
    ### Fill missing dates with NaN ###
    # Create a date range from  min to max
    dr = pd.date_range(start=x['day'].min(), end=max_date, freq='1D')
    # Add hh_0 to hh_47 to columns (for the 48 half-hourly measurements)
    dr = pd.DataFrame(columns=[f"hh_{i}" for i in range(48)], index=dr).unstack().reset_index()
    # renaming the columns
    dr.columns = ['hour_block', 'day', '_']
    # left merging the dataframe to the standard dataframe
    dr = dr.merge(x, on=['hour_block','day'], how='left') # missing values will be left as NaN
    # sorting the rows
    dr.sort_values(['day', 'offset'], inplace=True)
    # extracting the timeseries array
    ts = dr['energy_consumption'].values
    len_ts = len(ts)
    return start_date, name, ts, len_ts

def load_process_block_compact(block_df, max_date, freq='30min', 
                               ts_identifier='series_name', value_name='series_value'):
    """Takes a DataFrame for one of the blocks, a frequency, a time series identifier and 
    a value name. Applies compact processing to each of the LCLids (households) in the block.
    Returns the data for that block, converted into compact form."""
    grps = block_df.groupby('LCLid')
    all_series = []
    all_start_dates = []
    all_names = []
    all_data = {}
    all_len = []
    for idx, df in tqdm(grps, leave=False):
        start_date, name, ts, len_ts = preprocess_compact(df, max_date)
        all_series.append(ts)
        all_start_dates.append(start_date)
        all_names.append(name)
        all_len.append(len_ts)

    all_data[ts_identifier] = all_names
    all_data['start_timestamp'] = all_start_dates
    all_data['frequency'] = freq
    all_data[value_name] = all_series
    all_data['series_length'] = all_len
    return pd.DataFrame(all_data)

So that we can combine the energy consumption data with the weather and bank holiday data, we combine all block data into single data frame. To do this, we:

1) create a standard DataFrame using the start date and global end date (`max_date`)
2) left merge the DataFrame for each LCLid onto the standard DataFrame, leaving missing values as `np.nan`

block_df_l = []

# wrapping in tqdm shows a smart progress meter as blocks are processed
# glob matches all .csv files in the block_data_path directory
for file in tqdm(sorted(list(block_data_path.glob('*.csv'))), desc='processing blocks...'):
    block_df = pd.read_csv(file, parse_dates=False)
    block_df['day'] = pd.to_datetime(block_df['day'], yearfirst=True)
    # take only from 2012-01-01
    block_df = block_df.loc[block_df['day']>='2012-01-01']
    # reshape the dataframe into long form with hour blocks along the rows
    block_df = block_df.set_index(['LCLid', 'day']).stack().reset_index().rename(columns={'level_2': 'hour_block', 0: 'energy_consumption'})
    # creating a numerical hourblock column
    block_df['offset'] = block_df['hour_block'].str.replace("hh_", "").astype(int)
    block_df_l.append(load_process_block_compact(block_df, max_date, freq='30min', ts_identifier='LCLid', value_name='energy_consumption'))

hhblock_df = pd.concat(block_df_l)
del block_df_l

hhblock_df.head()

`informations_households.csv` contains metadata describing each household. Because these are not time-dependent (static) we left merge these on LCLid:

household_info = pd.read_csv(source_data/'informations_households.csv')
household_info.head()

hhblock_df = hhblock_df.merge(household_info, on='LCLid', validate='one_to_one')

hhblock_df.head()

`uk_bank_holidays.csv` describes dates of uk holidays. Energy consumption patterns might differ when households are spending more time in the home.

bank_holidays = pd.read_csv(source_data/'uk_bank_holidays.csv', parse_dates=False)
bank_holidays['Bank holidays'] = pd.to_datetime(bank_holidays['Bank holidays'], yearfirst=True)
bank_holidays.set_index('Bank holidays', inplace=True)
bank_holidays.head()

# reindex on standard date range
# resample the holiday data to every 30 minutes
# forward fill holidays within a day, fill NaN values with NO_HOLIDAY
bank_holidays = bank_holidays.resample('30min').asfreq()
bank_holidays = bank_holidays.groupby(bank_holidays.index.date).ffill().fillna('NO_HOLIDAY')
bank_holidays.index.name='datetime'
bank_holidays.head()

`weather_hourly_darksky.csv` contains the hourly data from the darksky weather API. We upsample this to every 30 minutes and forward-fill missing values.

weather_hourly = pd.read_csv(source_data/'weather_hourly_darksky.csv', parse_dates=False)
weather_hourly['time'] = pd.to_datetime(weather_hourly['time'], yearfirst=True)
weather_hourly.set_index('time', inplace=True)
weather_hourly.head()

# resample the data to every 30 minutes, forward fill
weather_hourly = weather_hourly.resample("30min").ffill()
weather_hourly.head()

Here, as a nicety, we define a function for mapping the relevant holiday and weather data onto each row so that we can use `progress_apply` to complete our DataFrame incrementally.

def map_holidays_weather(row, bank_holidays, weather_hourly):
    """Takes a row from the (compact) energy dataset, a DataFrame of bank holidays and
    a DataFrame of hourly weather data.
    Returns the row with the appropriate weather and energy data joined on."""
    date_range = pd.date_range(row['start_timestamp'], 
                               periods=row['series_length'], 
                               freq=row['frequency'])
    std_df = pd.DataFrame(index=date_range)
    # we fill NA values with NO_HOLIDAY because rows before an earlier holiday will be NaN
    holidays = std_df.join(bank_holidays, how='left').fillna('NO_HOLIDAY')
    weather = std_df.join(weather_hourly, how='left')
    assert len(holidays) == row['series_length'], 'Length of holidays should be same as series length'
    assert len(weather)== row['series_length'], 'Length of weather should be same as series length'
    row['holidays'] = holidays['Type'].values
    for col in weather:
        row[col] = weather[col].values
    return row

hhblock_df = hhblock_df.progress_apply(map_holidays_weather, args=(bank_holidays, weather_hourly), axis=1)
hhblock_df.head()

### Saving the combined data to disk

Split the DataFrame into chunks and save these chunks as parquet files. This has the advantages:

1. leverages the compression that comes with the format
2. reading in parts of the whole data allows for quick iteration and experimentation
3. data types are retained between read and write operations

os.makedirs('../data/preprocessed', exist_ok=True)

We save a map of which households belong to each Acorn group as a pickle file. This will help us read in households for each of these groups straight from our Parquet files later on.

# save the LCLid to Acorn group map as a pickle to be used later
hhblock_df[['LCLid', 'file', 'Acorn_grouped']].to_pickle(f'../data/preprocessed/london_smart_meters_lclid_acorn_map.pkl')

# split the blocks into 8 chunks
blocks = [f'block_{i}' for i in range(111)]

n_chunks= 8
split_blocks = [blocks[i:i + n_chunks] for i in range(0, len(blocks), n_chunks)] 

#Writing each chunk to disk
for blk in tqdm(split_blocks):
    df = hhblock_df.loc[hhblock_df.file.isin(blk)]
    blk = [int(b.replace('block_', '')) for b in blk]
    block_str = f'block_{min(blk)}-{max(blk)}'
    df.to_parquet(f'../data/preprocessed/london_smart_meters_merged_{block_str}.parquet')
