==> code/02-train-test-eval-split.ipynb <==
import numpy as np
import os
import pandas as pd
from collections import defaultdict
from pathlib import Path
from tqdm import tqdm
import statsmodels.api as sm
import warnings
import random
from IPython.display import display, HTML
from statsmodels.tsa.seasonal import seasonal_decompose
from matplotlib import pyplot as plt
np.random.seed(42)
random.seed(42)
tqdm.pandas()

## 02 - Creating Train, Test and Eval Splits

In this notebook, we create training, testing and evaluation splits of the data for building models.

We take 50 households from each of the three Acorn groups for now, in order to reduce the burden on our hardware (peforming forecasts with large numbers of series may be memory intensive). Note that this means we're ignoring households with an unknown Acorn group.

# Read in the map of acord household types
lclid_acorn_map = pd.read_pickle('../data/preprocessed/london_smart_meters_lclid_acorn_map.pkl')

# Split the households according to their acorn group
affluent_households = lclid_acorn_map.loc[lclid_acorn_map.Acorn_grouped=='Affluent', ['LCLid', 'file']]
adversity_households = lclid_acorn_map.loc[lclid_acorn_map.Acorn_grouped=='Adversity', ['LCLid', 'file']]
comfortable_households = lclid_acorn_map.loc[lclid_acorn_map.Acorn_grouped=='Comfortable', ['LCLid', 'file']]

households_per_acorn = 50
random_state = 76 #  Set for reproducibility

selected_households = pd.concat(
    [
        affluent_households.sample(households_per_acorn, random_state=random_state),
        comfortable_households.sample(households_per_acorn, random_state=random_state),
        adversity_households.sample(households_per_acorn, random_state=random_state),
    ]
)
selected_households['block'] = selected_households.file.str.split('_', expand=True).iloc[:,1].astype(int)

# Extract the paths to the different blocks, extracting the starting and ending blocks
path_blocks = [
    (p, *list(map(int, p.name.split('_')[5].split('.')[0].split('-'))))
    for p in Path('../data/preprocessed').glob(
        'london_smart_meters_merged_block*'
    )
]

household_df_l = []
for path, start_b, end_b in tqdm(path_blocks):
    block_df = pd.read_parquet(path)
    selected_households['block'].between
    mask = selected_households['block'].between(start_b, end_b)
    lclids = selected_households.loc[mask, "LCLid"]
    household_df_l.append(block_df.loc[block_df.LCLid.isin(lclids)])

block_df = pd.concat(household_df_l)
del household_df_l
block_df.head()

We need to convert the compact form of the time series columns into the expanded form to make them easier to work with.

First define a function for converted each time series from its compact form to expanded form:

def compact_to_expanded(df, timeseries_col, static_cols, time_varying_cols, ts_identifier):
    """Takes a DataFrame of time series, a time series column, a list of static columns,
    a list of time-varying columns and an identifier column name. 
    Returns an expanded form version of the DataFrame."""

    def preprocess_expanded(x):
        # Create a date range from start
        dr = pd.date_range(
            start = x['start_timestamp'],
            periods = len(x['energy_consumption']),
            freq = x['frequency'],
        )
        # Build a dictionary of lists for each of our columns,
        # then convert this into a DataFrame for that time series
        df_columns = defaultdict(list)
        df_columns['timestamp'] = dr
        for col in [ts_identifier, timeseries_col] + static_cols + time_varying_cols:
            df_columns[col] = x[col]
        return pd.DataFrame(df_columns)

    # expand each time series row, then join these expansions together into a single DataFrame
    all_series = []
    for i in tqdm(range(len(df))):
        all_series.append(preprocess_expanded(df.iloc[i]))
    df = pd.concat(all_series)
    del all_series
    return df

#convert our time series to expanded form
exp_block_df = compact_to_expanded(block_df, timeseries_col = 'energy_consumption',
static_cols = ["frequency", "series_length", "stdorToU", "Acorn", "Acorn_grouped", "file"],
time_varying_cols = ['holidays', 'visibility', 'windBearing', 'temperature', 'dewPoint',
       'pressure', 'apparentTemperature', 'windSpeed', 'precipType', 'icon',
       'humidity', 'summary'],
ts_identifier = "LCLid")

exp_block_df.head()

exp_block_df.info(memory_usage="deep", verbose=False)

We can reduce the memory footprint of the expanded data by converting the datatypes of particular columns:

def reduce_memory_footprint(df):
    dtypes = df.dtypes
    object_cols = dtypes[dtypes == "object"].index.tolist()
    float_cols = dtypes[dtypes == "float64"].index.tolist()
    int_cols = dtypes[dtypes == "int64"].index.tolist()
    df[int_cols] = df[int_cols].astype("int32")
    df[object_cols] = df[object_cols].astype("category")
    df[float_cols] = df[float_cols].astype("float32")
    return df

exp_block_df = reduce_memory_footprint(exp_block_df)

exp_block_df.info(memory_usage="deep", verbose=False)

We have data from the beginning of November 2011 until the end of February 2014. 

We will set aside data from 2014 as our validation and testing data, with January as validation and February as test. 

test_mask = (exp_block_df.timestamp.dt.year == 2014) & (exp_block_df.timestamp.dt.month == 2)
val_mask = (exp_block_df.timestamp.dt.year == 2014) & (exp_block_df.timestamp.dt.month == 1)

train = exp_block_df[~(val_mask|test_mask)]
val = exp_block_df[val_mask]
test = exp_block_df[test_mask]
print(f'# of Training samples: {len(train)} | # of Validation samples: {len(val)} | # of Test samples: {len(test)}')
print(f'Max Date in Train: {train.timestamp.max()} | Min Date in Validation: {val.timestamp.min()} | Min Date in Test: {test.timestamp.min()}')

### Filling missing values with seasonal interpolation

Before filling in any missing values in the energy consumption series using imputation, we save the raw (untouched) version of our train, val and test sets to parquet:

train.to_parquet('../data/preprocessed/selected_blocks_train.parquet')
val.to_parquet('../data/preprocessed/selected_blocks_val.parquet')
test.to_parquet('../data/preprocessed/selected_blocks_test.parquet')

def seasonal_interpolation(data):
    """Takes a time series and returns a seasonal-interpolated version
    of the time series."""
    res = data.copy()
    # Seasonal_decompose doesn't work where the series has missing values,
    # so for now we interpolate these (using limit_direction = 'both')

    na_index = pd.isna(data)
    temp = data.interpolate(method='spline', order=3, limit_direction = 'both')
    dr = seasonal_decompose(temp, period=48*7)
    # join trend and irregular component (timeseries without seasonality):
    data_no_seasonality = dr.trend + dr.resid
    # fill in NA values:
    data_no_seasonality[na_index] = np.nan
    # interpolate data without seasonality:
    data_no_seasonality_imputed = data_no_seasonality.interpolate(method='spline', order=3, limit_direction = 'both')
    # add back seasonality:
    data_imputed = data_no_seasonality_imputed + dr.seasonal
    # merge interpolated values into original timeseries:
    res[na_index] = data_imputed[na_index]

    return res

# apply seasonal interpolation to each household 

for lclid in tqdm(selected_households.LCLid):
    ts_df = exp_block_df[exp_block_df.LCLid == lclid].set_index('timestamp')
    ts_interpolated = seasonal_interpolation(ts_df.energy_consumption).reset_index()
    exp_block_df.loc[(exp_block_df['LCLid'] == lclid) & (exp_block_df['timestamp'].isin(ts_interpolated['timestamp'])), 
    'energy_consumption'] = ts_interpolated['energy_consumption']

Here, we create train, val and test sets after applying seasonal interpolation to fill in missing values.

train = exp_block_df[~(val_mask|test_mask)]
val = exp_block_df[val_mask]
test = exp_block_df[test_mask]
print(f'# of training samples: {len(train)} | # of validation samples: {len(val)} | # of Test samples: {len(test)}')
print(f'Max Date in train: {train.timestamp.max()} | Min Date in validation: {val.timestamp.min()} | Min Date in test: {test.timestamp.min()}')

train.to_parquet('../data/preprocessed/selected_blocks_train_missing_imputed.parquet')
val.to_parquet('../data/preprocessed/selected_blocks_val_missing_imputed.parquet')
test.to_parquet('../data/preprocessed/selected_blocks_test_missing_imputed.parquet')
